helpers/detection.py
"""
YOLO Detection Module for Clean City Waste Detection
Handles YOLOv8 model loading, inference, and waste classification
"""

import cv2
import numpy as np
from ultralytics import YOLO
import torch
from typing import List, Dict, Tuple, Optional
import os
import streamlit as st

class WasteDetector:
    """
    YOLOv8-based waste detection system for identifying plastic, paper, and organic waste
    """
    
    def __init__(self, model_path: str = None):
        """
        Initialize the waste detector with YOLOv8 model
        
        Args:
            model_path: Path to YOLOv8 weights file. If None, uses default YOLOv8n model
        """
        self.class_names = {
            0: 'plastic',
            1: 'paper', 
            2: 'organic'
        }
        
        # Color mapping for each waste type (BGR format for OpenCV)
        self.class_colors = {
            'plastic': (255, 0, 0),    # Blue
            'paper': (0, 255, 255),    # Yellow
            'organic': (0, 255, 0)     # Green
        }
        
        self.confidence_threshold = 0.25
        self.iou_threshold = 0.45
        
        # Load YOLO model
        self.model = self._load_model(model_path)
        
    def _load_model(self, model_path: Optional[str]) -> YOLO:
        """
        Load YOLOv8 model from path or use pretrained model
        
        Args:
            model_path: Path to custom model weights
            
        Returns:
            Loaded YOLO model
        """
        try:
            # Check for trained model first
            trained_model_path = "yolo_model/yolov8_waste_trained.pt"
            if os.path.exists(trained_model_path):
                print(f"Loading trained model from: {trained_model_path}")
                model = YOLO(trained_model_path)
                st.success(f"✅ Loaded trained waste detection model (97.1% mAP)")
                print("Trained model loaded successfully")
                return model
            elif model_path and os.path.exists(model_path):
                # Load custom trained model
                model = YOLO(model_path)
                st.success(f"✅ Loaded custom model from {model_path}")
                return model
            else:
                # Use pretrained YOLOv8n model and fine-tune for waste detection
                model = YOLO('yolov8n.pt')
                st.info("📦 Using pretrained YOLOv8n model (will detect general objects)")
                return model
                
        except Exception as e:
            st.error(f"❌ Error loading model: {str(e)}")
            # Fallback to basic YOLOv8n
            return YOLO('yolov8n.pt')
    
    def detect_waste(self, image: np.ndarray) -> Dict:
        """
        Detect waste objects in an image
        
        Args:
            image: Input image as numpy array (BGR format)
            
        Returns:
            Dictionary containing detection results
        """
        try:
            # Run YOLO inference
            results = self.model(image, 
                               conf=self.confidence_threshold,
                               iou=self.iou_threshold,
                               verbose=False)
            
            detections = []
            class_counts = {'plastic': 0, 'paper': 0, 'organic': 0}
            
            # Process detection results
            if results and len(results) > 0:
                result = results[0]
                
                if result.boxes is not None and len(result.boxes) > 0:
                    boxes = result.boxes.xyxy.cpu().numpy()
                    confidences = result.boxes.conf.cpu().numpy()
                    class_ids = result.boxes.cls.cpu().numpy().astype(int)
                    
                    for i, (box, conf, class_id) in enumerate(zip(boxes, confidences, class_ids)):
                        # Map YOLO classes to waste categories
                        waste_class = self._map_yolo_to_waste_class(class_id)
                        
                        # For trained model, process all detections; for pretrained, only waste-related
                        trained_model_exists = os.path.exists("yolo_model/yolov8_waste_trained.pt")
                        if waste_class:
                            x1, y1, x2, y2 = box.astype(int)
                            
                            detection = {
                                'bbox': [x1, y1, x2, y2],
                                'confidence': float(conf),
                                'class': waste_class,
                                'color': self.class_colors[waste_class]
                            }
                            
                            detections.append(detection)
                            class_counts[waste_class] += 1
            
            return {
                'detections': detections,
                'class_counts': class_counts,
                'total_objects': len(detections)
            }
            
        except Exception as e:
            st.error(f"❌ Detection error: {str(e)}")
            return {
                'detections': [],
                'class_counts': {'plastic': 0, 'paper': 0, 'organic': 0},
                'total_objects': 0
            }
    
    def _map_yolo_to_waste_class(self, yolo_class_id: int) -> Optional[str]:
        """
        Map YOLO class IDs to waste categories
        
        Args:
            yolo_class_id: YOLO detected class ID
            
        Returns:
            Waste category name or None if not waste-related
        """
        # Check if we're using the trained model (custom classes)
        trained_model_path = "yolo_model/yolov8_waste_trained.pt"
        if os.path.exists(trained_model_path):
            # Custom trained model class mapping
            waste_mapping = {
                0: 'plastic',   # Class 0 in trained model
                1: 'paper',     # Class 1 in trained model  
                2: 'organic'    # Class 2 in trained model
            }
            return waste_mapping.get(yolo_class_id, None)
        else:
            # YOLO COCO class mappings to waste categories (for pretrained model)
            waste_mapping = {
                # Plastic items
                39: 'plastic',  # bottle
                40: 'plastic',  # wine glass
                41: 'plastic',  # cup
                42: 'plastic',  # fork
                43: 'plastic',  # knife
                44: 'plastic',  # spoon
                45: 'plastic',  # bowl
                
                # Paper items  
                73: 'paper',    # book
                74: 'paper',    # clock (assuming paper/cardboard)
                
                # Organic items
                46: 'organic',  # banana
                47: 'organic',  # apple
                48: 'organic',  # sandwich
                49: 'organic',  # orange
                50: 'organic',  # broccoli
                51: 'organic',  # carrot
                52: 'organic',  # hot dog
                53: 'organic',  # pizza
                54: 'organic',  # donut
                55: 'organic',  # cake
            }
            return waste_mapping.get(yolo_class_id, None)
    
    def detect_waste_in_video(self, video_path: str, sample_rate: int = 30) -> List[Dict]:
        """
        Detect waste in video by sampling frames
        
        Args:
            video_path: Path to video file
            sample_rate: Process every Nth frame
            
        Returns:
            List of detection results for sampled frames
        """
        try:
            cap = cv2.VideoCapture(video_path)
            frame_results = []
            frame_count = 0
            
            while cap.isOpened():
                ret, frame = cap.read()
                if not ret:
                    break
                
                # Sample frames at specified rate
                if frame_count % sample_rate == 0:
                    result = self.detect_waste(frame)
                    result['frame_number'] = frame_count
                    frame_results.append(result)
                
                frame_count += 1
            
            cap.release()
            return frame_results
            
        except Exception as e:
            st.error(f"❌ Video processing error: {str(e)}")
            return []
    
    def update_confidence_threshold(self, threshold: float):
        """Update confidence threshold for detections"""
        self.confidence_threshold = max(0.1, min(1.0, threshold))
    
    def update_iou_threshold(self, threshold: float):
        """Update IoU threshold for NMS"""
        self.iou_threshold = max(0.1, min(1.0, threshold))


def create_demo_detector() -> WasteDetector:
    """
    Create a demo waste detector instance
    
    Returns:
        Configured WasteDetector instance
    """
    model_path = os.path.join("yolo_model", "yolov8_weights.pt")
    return WasteDetector(model_path)


def simulate_waste_detection(image: np.ndarray) -> Dict:
    """
    Simulate waste detection for demo purposes when no trained model is available
    
    Args:
        image: Input image
        
    Returns:
        Simulated detection results
    """
    height, width = image.shape[:2]
    
    # Generate some random detections for demo
    np.random.seed(42)  # For consistent demo results
    
    detections = []
    class_counts = {'plastic': 0, 'paper': 0, 'organic': 0}
    
    # Simulate 2-5 random detections
    num_detections = np.random.randint(2, 6)
    
    for _ in range(num_detections):
        # Random bounding box
        x1 = np.random.randint(0, width // 2)
        y1 = np.random.randint(0, height // 2)
        x2 = np.random.randint(x1 + 50, min(x1 + 200, width))
        y2 = np.random.randint(y1 + 50, min(y1 + 200, height))
        
        # Random class
        waste_class = np.random.choice(['plastic', 'paper', 'organic'])
        confidence = np.random.uniform(0.6, 0.95)
        
        colors = {
            'plastic': (255, 0, 0),    # Blue
            'paper': (0, 255, 255),    # Yellow
            'organic': (0, 255, 0)     # Green
        }
        
        detection = {
            'bbox': [x1, y1, x2, y2],
            'confidence': confidence,
            'class': waste_class,
            'color': colors[waste_class]
        }
        
        detections.append(detection)
        class_counts[waste_class] += 1
    
    return {
        'detections': detections,
        'class_counts': class_counts,
        'total_objects': len(detections)
    }
